% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeAllCaseControl_v1.5.R
\name{CreateObject_fromSeurat}
\alias{CreateObject_fromSeurat}
\title{Create Nichobject from Seurat object}
\usage{
CreateObject_fromSeurat(
  Seurat.object,
  slot = "counts",
  status = NULL,
  expDirection = "BOTH",
  cell_type = "orig.ident",
  data_source = "UMI",
  scale.factor = 10^6,
  Org = "Homo sapiens"
)
}
\arguments{
\item{Seurat.object}{The Seurat object which stores the expression matrix and cell type information}

\item{slot}{The name of slot which contains expression matrix, default "counts".}

\item{status}{the status of cell disease or control}

\item{expDirection}{the direction of expression change UP DOWN or BOTH}

\item{cell_type}{The name of specific column which contains cell type information, default "orig.ident".}

\item{scale.factor}{set the scale factor, default "10^6"}

\item{Org}{the species of this scRNA-seq}

\item{source}{the type of expression dataframe, eg "UMI", "fullLength", "TPM", or "CPM"}
}
\description{
Create Nichobject from Seurat object
}
