% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeAllCaseControl_v1.5.R
\name{getForBubble}
\alias{getForBubble}
\title{melt enrich dataframe for bubble graph}
\usage{
getForBubble(
  pathway.hyper.list = pathway.hyper.list,
  cella_cellb,
  Org = "Homo sapiens"
)
}
\arguments{
\item{pathway.hyper.list}{a list of pathway enrichment pvalue and NES}

\item{Org}{the species of this scRNA-seq}
}
\value{
the result for pbubble graph
}
\description{
melt enrich dataframe for bubble graph
}
