% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeAllCaseControl_v1.5.R
\name{mylog2foldChange.diy.casecontrol}
\alias{mylog2foldChange.diy.casecontrol}
\title{calculate foldchange between case and control for each celltype}
\usage{
mylog2foldChange.diy.casecontrol(
  inData,
  cell.type,
  method = "median",
  probs = 0.75,
  status = NULL
)
}
\arguments{
\item{inData}{a dataframe of gene expression}

\item{cell.type}{the cell type which you want to calculate the foldchange}

\item{method}{default is "median",eg "median" or "mean"}

\item{probs}{the quantile for median calculate}
}
\value{
the foldchange value of each celltype in \code{cellwave object}
}
\description{
calculate foldchange between case and control for each celltype
}
